requirements.txt
streamlit
langchain
openai
gitpython
faiss-cpu
langchain-community
tiktoken
langchain-openai